requirements.txt
pyyaml
torchvision
torch>=1.8.0
tqdm
numpy
loguru
opencv-python
matplotlib
xformer
tabulate